Ignore unused class variable, unused argument and return value discarded arguments
The unused class variable warning triggers even if the class variable is used elsewhere. Sometimes a function will have an argument that is unused in order to be able to be linked with a signal. And functions like connect() will often have their return value discarded; checking each call of that seems like too much.